Lesson-6 add parameterized tests

+++ bus-tickets/src/main/java/halatsiankova/javafromscratch/busTicket/service/TicketService.java
@@ -19,28 +19,42 @@ public class TicketService {
         this.generateUUID = generateId;
     }
 
-    public void save(BusTicket ticket) {
-        ticket.setId(generateUUID.generateId());
-        repository.save(ticket);
-    }
-
-    public Collection<BusTicket> saveAll(Collection<BusTicket> collection) {
-        collection.forEach(ticket -> ticket.setId(generateUUID.generateId()));
-        repository.saveAll(collection);
-        return collection;
+    /**
+     * Saves a collection of BusTicket objects by first clearing the existing tickets in the repository,
+     * then assigning a unique UUID to each new ticket, and finally saving all the tickets in the repository.
+     * @param tickets the collection of BusTicket objects to save
+     * @return the collection of BusTicket objects after they have been saved, with their IDs set
+     */
+    public Collection<BusTicket> saveAll(Collection<BusTicket> tickets) {
+        if(tickets == null) {
+            throw new IllegalArgumentException("Collection of tickets must not be null.");
+        }
+        repository.deleteAll();
+        tickets.forEach(ticket -> ticket.setId(generateUUID.generateId()));
+        repository.saveAll(tickets);
+        return tickets;
     }
 
     public BusTicket getById(UUID ticketId) {
+        if(ticketId == null) {
+            throw new IllegalArgumentException("Ticket ID must not be null.");
+        }
         return repository.findById(ticketId).
                 orElseThrow(() -> new IllegalArgumentException(String.format("BusTicket with id=%s doesn't exist.",
                         ticketId)));
     }
 
     public void deleteByTicketId(UUID ticketId) {
+        if(ticketId == null) {
+            throw new IllegalArgumentException("Ticket ID must not be null.");
+        }
         repository.deleteById(ticketId);
     }
 
     public List<BusTicket> getAllByTicketIdAndPriceFromTo(TicketType ticketType, BigDecimal priceFrom, BigDecimal priceTo) {
+        if(ticketType == null || priceFrom == null || priceTo == null) {
+            throw new IllegalArgumentException("Ticket type, price from, price to must not be null.");
+        }
         if(priceFrom.compareTo(priceTo) >0) {
             throw new IllegalArgumentException("The price from must be less than the price to.");
         }
